+++ domain/src/main/java/com/blogspot/soyamr/domain/models/Geolocation.kt
@@ -1,6 +1,6 @@
 package com.blogspot.soyamr.domain.models
 
 data class Geolocation(
-    val lat: Long,
-    val lng: Long,
+    val lat: Double,
+    val lng: Double,
 )
